feat: allow users to choose a group to share with
